FASE III START
Bastantes cosas:

CONTROL DE ACCESO :anguished:
Modificado las entidades
Creado el control de acceso por roles funcional
AUTENTICACIÓN :100:
Protección CSRF
Creado el autenticador
CONFIGURACIONES :coffee:
Parte web hecha, falta acabar de securizar API REST
FILTROS :sob:
Movidón lo de los @JsonViews e o no flipas al final he quitao todo jsjs
SERVICIOS :rage1:
Añadido par de funcionalidades a los servicios
JAVASCRIPT :+1:
He creado unos JS para la gestión de entidades desde la vista Admin, pero falta por integrar :()
CONTROLLERS :snail:
Puf no veas que pereza pa mapear toda la pagina
HTML :-1:
He movido todos los estaticos a templates y me he creado un head para importarlo en los templates y asi evitarme el añadir el token csrf a todos los templates

+++ src/main/java/com/example/PracticaWeb/Service/TicketService.java
@@ -18,7 +18,7 @@ public class TicketService {
     TicketRepository ticketRepository;
 
     //CHECK IF EXISTS
-    public Optional<Ticket> existsById(Integer id){
+    public Optional<Ticket> existsById(long id){
         var aux = ticketRepository.findTicketByid(id);
         return aux;
 
@@ -41,7 +41,7 @@ public class TicketService {
         return ticketRepository.save(ticket);
 
     }
-    public Optional<Ticket> deleteTicket(Integer id){
+    public Optional<Ticket> deleteTicket(long id){
         Optional<Ticket> aux = ticketRepository.findTicketByid(id);
         if(aux.isPresent()){
             ticketRepository.delete(aux.get());
@@ -49,7 +49,7 @@ public class TicketService {
         return aux;
 
     }
-    public Optional<Ticket> returnTicket(Integer id){
+    public Optional<Ticket> returnTicket(long id){
         return ticketRepository.findTicketByid(id);
 
     }
